// src/client/types/table-design.ts
/** 整页 */
export const PAGEALL = 'page-all'

/** 左右布局 */
export const ABOUTLAYOUT = 'about-layout'

/** 整页 tabs 布局 */
export const PAGEALLTABS = 'page-all-tabs'

/** 左右 tabs 布局 */
export const ABOUTLAYOUTTABS = 'about-layout-tabs'

/** 控件类型 */
export enum CONTROL {
  /** proTable */
  TABLE = 'table',
  /** tabs */
  TABS = 'tabs',

  TREE = 'tree',

  COLUMNS = 'columns',
  OTHERATTR = 'otherAttr',
  REQUEST = 'request',
  TOOLBARRENDER = 'toolBarRender',
  HEADERTITLE = 'headerTitle',
  TREEATTR = 'treeAttr',
  TABSATTR = 'tabsAttr'
  // COLUMNS = 'columns',
  // COLUMNS = 'columns',
  // COLUMNS = 'columns',
  // COLUMNS = 'columns',
}